requirements.txt
paho-mqtt>=2.0.0
aiosqlite>=0.18.0
flask>=3.0.0
